requirements.txt
fastapi
uvicorn[standard]
langchain-core
langchain-openai
langchain-text-splitters
langchain-pinecone
langgraph
pydantic
numpy
python-dotenv
python-docx
firebase-admin
python-multipart
email-validator
pandas
openpyxl
chardet
httpx
resend
